pyproject.toml
[project]
name = "pilot-dub"
version = "0.1.0"
description = "PILOT-Dub visual TTS generator-reranker framework"
readme = "README.md"
requires-python = ">=3.10,<3.11" # 推荐 3.10，兼顾兼容性与性能

dependencies = [
    "bitarray>=2.9.0",
    "dlib>=19.24.0",
    "einops>=0.8.0",
    "encodec>=0.1.1",
    "flashy>=0.0.2",
    "hydra-core>=1.3.2",
    "joblib>=1.4.0",
    "librosa>=0.10.0",
    "lightning>=2.4.0",      # 升级到较新版本
    "numpy==1.23.5",         # 现代版本，支持新版特性
    "omegaconf>=2.3.0",
    "opencv-python-headless>=4.10.0.0", # 服务器端推荐用 headless 避免 GUI 报错
    "phonemizer>=3.2.1",
    "pydub>=0.25.1",
    "python_speech_features>=0.6",
    "sacrebleu>=2.4.0",
    "scikit-image>=0.24.0",
    "scikit-learn>=1.5.0",
    "scikit-video>=1.1.11",
    "scipy>=1.13.0",
    "sentencepiece>=0.2.0",
    "soundfile>=0.12.1",
    "submitit>=1.5.3",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.66.0",
    "transformers>=4.46.0",  # 升级到目前主流稳定版
    "wandb>=0.18.0",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.pilot_dub]
extra-pythonpath = ["fairseq", "avhubert"]

[tool.pilot_dub.conda_dependencies]
# 针对 4090 环境，建议在新机器上先执行：
# conda install -c conda-forge icu=73.2 libsqlite=3.45.0
icu = "==73.2"
libsqlite = "==3.45.0"
